+++ apollo-client/src/main/java/com/ctrip/framework/apollo/spi/DefaultConfigFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Apollo Authors
+ * Copyright 2022 Apollo Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -76,7 +76,7 @@ public class DefaultConfigFactory implements ConfigFactory {
         format != ConfigFileFormat.Properties) {
       configRepository = createPropertiesCompatibleFileConfigRepository(namespace, format);
     } else {
-      configRepository = createLocalConfigRepository(namespace);
+      configRepository = createConfigRepository(namespace);
     }
 
     logger.debug("Created a configuration repository of type [{}] for namespace [{}]",
@@ -91,7 +91,7 @@ public class DefaultConfigFactory implements ConfigFactory {
 
   @Override
   public ConfigFile createConfigFile(String namespace, ConfigFileFormat configFileFormat) {
-    ConfigRepository configRepository = createLocalConfigRepository(namespace);
+    ConfigRepository configRepository = createConfigRepository(namespace);
     switch (configFileFormat) {
       case Properties:
         return new PropertiesConfigFile(namespace, configRepository);
@@ -108,6 +108,13 @@ public class DefaultConfigFactory implements ConfigFactory {
     }
 
     return null;
+  }
+
+  ConfigRepository createConfigRepository(String namespace) {
+    if (m_configUtil.isPropertyFileCacheEnabled()) {
+      return createLocalConfigRepository(namespace);
+    }
+    return createRemoteConfigRepository(namespace);
   }
 
   /**