+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/portal/service/AppService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Apollo Authors
+ * Copyright 2022 Apollo Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -161,13 +161,6 @@ public class AppService {
     if (managedApp != null) {
       return app;
     }
-
-    UserInfo owner = userService.findByUserId(app.getOwnerName());
-    if (owner == null) {
-      throw new BadRequestException("Application's owner not exist.");
-    }
-
-    app.setOwnerEmail(owner.getEmail());
 
     App createdApp = appRepository.save(app);